Grow EPA crawl again
The cause of the weekend's problems was that a bunch of EPA subdomains had gone offline, causing extremely long timeouts and delays. Splitting the EPA crawl up into more pieces was a decent quick fix, but ideally we probably don't want that much parallelism (trying to be nice to the servers, but also each parallel job is more duplicated subresource data in our WARCs).

All the offline subdomains were fixed today, so this is no longer an *active* problem. A better long-term fix is the `--precheck-connections` option (currently sitting in a PR).