Import precheck errors as records in DB
We've previously been recording these manually when especially relevant, but that's caused us to miss some important data and not keep up with things effectively. I was originally worried about recording lots of erroneous errors, but at this point I think we're better recording them all than failing in the way we currently are.